requirements.txt
Flask==0.10.1
Flask-login==0.4.1
Flask-SQLAlchemy==2.1
PyMySQL==0.7.9
PyJWT==1.4.0
bcrypt==3.1.2
